relocated all logic from controllers to services

+++ src/main/java/com/nic/newspaper/service/CommentServiceImpl.java
@@ -1,12 +1,17 @@
 package com.nic.newspaper.service;
 
+import java.text.ParseException;
 import java.text.SimpleDateFormat;
+import java.util.Comparator;
 import java.util.Date;
 import java.util.List;
 
+import org.apache.commons.logging.Log;
+import org.apache.commons.logging.LogFactory;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
+import com.nic.newspaper.config.JwtTokenUtil;
 import com.nic.newspaper.dao.ArticleDao;
 import com.nic.newspaper.dao.CommentDao;
 import com.nic.newspaper.dao.UserDAO;
@@ -14,25 +19,56 @@ import com.nic.newspaper.entity.Article;
 import com.nic.newspaper.entity.Comment;
 import com.nic.newspaper.model.CrmComment;
 
+import io.jsonwebtoken.ExpiredJwtException;
+import jakarta.servlet.http.HttpServletRequest;
 import jakarta.transaction.Transactional;
 
 @Service
 public class CommentServiceImpl implements CommentService {
 
 	@Autowired
+	ArticleService articleService;
+	
+	@Autowired
 	public ArticleDao articleDao;
 
 	@Autowired
 	private UserDAO userDao;
+	
+	@Autowired
+	private CommentService commentService;
 
 	@Autowired
 	private CommentDao commentDao;
+	
+	@Autowired
+	private JwtTokenUtil jwtTokenUtil;
+	
+	protected final Log logger = LogFactory.getLog(getClass());
 
 	SimpleDateFormat formatter = new SimpleDateFormat("dd.MM.yyyy HH:mm:ss");
 
 	@Override
 	@Transactional
-	public Article save(CrmComment theComment, String userEmail, long articleId) {
+	public Article save(long articleId, CrmComment theComment, HttpServletRequest request) {
+		
+		final String requestTokenHeader = request.getHeader("Authorization");
+
+		String userEmail = null;
+		String jwtToken = null;
+
+		if (requestTokenHeader != null && requestTokenHeader.startsWith("Bearer ")) {
+			jwtToken = requestTokenHeader.substring(7);
+			try {
+				userEmail = jwtTokenUtil.getUserEmailFromToken(jwtToken);
+			} catch (IllegalArgumentException e) {
+				System.out.println("Unable to get JWT Token");
+			} catch (ExpiredJwtException e) {
+				System.out.println("JWT Token has expired");
+			}
+		} else {
+			logger.warn("JWT Token is null or does not begin with Bearer String");
+		}
 
 		Article theArticle = articleDao.findArticleById(articleId);
 
@@ -66,17 +102,41 @@ public class CommentServiceImpl implements CommentService {
 
 	@Override
 	@Transactional
+	public void deleteCommentById(long commentId) {
+		
+		Comment tempComment = commentService.findCommentById(commentId);
+
+		if (tempComment == null) {
+			throw new RuntimeException("Comment id not found - " + commentId);
+		}
+
+		commentDao.deleteCommentById(commentId);
+
+	}
+
+	@Override
+	@Transactional
 	public Comment findCommentById(long commentId) {
 
 		return commentDao.findCommentById(commentId);
 	}
 
 	@Override
-	@Transactional
-	public void deleteCommentById(long commentId) {
+	public List<Comment> aritcleComments(long aritcleId) {
+		
+		List<Comment> comments = articleService.findArticleById(aritcleId).getComments();
 
-		commentDao.deleteCommentById(commentId);
+		Comparator<Comment> byDate = (first, second) -> {
+			try {
+				return formatter.parse(second.getDate()).compareTo(formatter.parse(first.getDate()));
+			} catch (ParseException e) {
+				e.printStackTrace();
+			}
+			return 0;
+		};
 
+		comments.sort(byDate);
+
+		return comments;
 	}
-
 }
